+++ CHANGELOG.md
@@ -57,3 +57,10 @@
 
 ## 3.0.0
 - add signature encoding
+
+## 3.0.1
+- add segwit type to signing
+
+
+## 3.0.1+1
+- add correct recId calculation